Isolate paths
`./parse_diagnostics.sh` was failing because the `./` refers to the client's
repository root. Use `${{ github.action_path }}` instead to refer to the
action repo root.

Introduce a temporary directory to download VisualSoar into, as well, just so
we don't pollute the client's repository with a copy of VisualSoar (what if they
decided to zip and upload that directory, for example?).