Don't upload for user; configurable output path
Uploading would fail if there were multiple projects being validated in
workflow. Even if that were not an issue, it's better to let the user decide
to upload the report.

Add an `outputPath` input and output for specifying where the output will be/was
stored. Then the user can upload that as an artifact if they want.